requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
Pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
scikit-learn==1.4.1.post1